% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_hss_clean_audit.R
\name{A_hss_clean_audit}
\alias{A_hss_clean_audit}
\title{Add duration columns for audit + filters}
\usage{
A_hss_clean_audit(dat)
}
\arguments{
\item{dat}{The resulting dataframe from hss_load_audit}
}
\description{
This fuction applis some basic cleaning (currently only adds a new column)
containing the duration spend on a screen.
}
