% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_hss_enum_assessment.R
\name{A_hss_enum_assessment}
\alias{A_hss_enum_assessment}
\title{Perform enumerator assessment on HSS dataframe}
\usage{
A_hss_enum_assessment(dat, audit = NULL)
}
\arguments{
\item{dat}{A dataframe containing the HSS survey data.}

\item{audit}{An audit file created with \code{hss_load_audit()}}
}
\value{
Returns a list containing an overview dataframe & a list of individual
dataframes for all criteria.
}
\description{
This function is used to perform the first step in the enumerator assessment.
It groups output by device ID and shows the total surveys collected and the
number of surveys remained after each of the selection criteria is applied. The
function accepts an optional argument of an audit dataframe created with
\code{hss_load_audit()}. If this argument is left empty the final selection criterion is skipped.
}
