% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_hss_id.R
\name{A_hss_id}
\alias{A_hss_id}
\title{Convert DeviceID to Name Reference}
\usage{
A_hss_id(dat, id, name)
}
\arguments{
\item{dat}{Dataframe (either audit or daily report)-oriented}

\item{id}{Numerical deviceid provided by ODK/KOBO}

\item{name}{Actual name of enumerator}
}
\value{
a modified dataframe that includes the actual names of enumerators.
}
\description{
This function is used in both the audit (enumerator assessmet) procedure,
and the daily report procedure. Numerical deviceids are converted to names.
This is a temporary function, as it still requires significant user
input. Improvements to this fucntion would involve reading pair-information
from an xls sheet.
}
