% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_hss_load_audit.R
\name{A_hss_load_audit}
\alias{A_hss_load_audit}
\title{Load audit files for HSS survey}
\usage{
A_hss_load_audit(path)
}
\arguments{
\item{path}{Path to a folder containing the audit .csv files}
}
\value{
data-frame with all audit files
}
\description{
This function is used to load individual audit files for a HSS survey as a
single dataframe.
}
