% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_hss_load_audit_KOBO.R
\name{A_hss_load_audit_KOBO}
\alias{A_hss_load_audit_KOBO}
\title{Load audit files for HSS survey (KOBO)}
\usage{
A_hss_load_audit_KOBO(path)
}
\arguments{
\item{path}{Path to a folder containing the audit .csv files}
}
\value{
data-frame with all audit files
}
\description{
This function is used to load individual audit files for a HSS survey as a
single dataframe.This is likely a redundant function as it relies on
KOBO.
}
