% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_hss_read_ODK.R
\name{A_hss_read_ODK}
\alias{A_hss_read_ODK}
\title{Load audit files for HSS survey (ODK)}
\usage{
A_hss_read_ODK(dat)
}
\arguments{
\item{dat}{Audit CSV produced by ODK export}
}
\value{
a dataframe that can be used to generate an audit report
}
\description{
This function is used to load individual audit files for a HSS survey as a
single dataframe.
}
