% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_tables.R
\name{append_tables}
\alias{append_tables}
\title{Append all data files stored in the same folder + file name}
\usage{
append_tables(
  source_folder,
  destination_folder,
  destination_file,
  file_pattern,
  file_separator
)
}
\arguments{
\item{source_folder}{The folder in which the to-be appended data files are stored}

\item{destination_folder}{The folder in which the appended data file should be stored}

\item{destination_file}{The name of the appended data file}

\item{file_pattern}{Set to ".csv" when appending CSV files}

\item{file_separator}{Set to "," when appending CSV files}
}
\value{
A CSV file containing the contents of all source files, incl. a new column containing the file names of the source files. NOTE: If the source files contain column headers, the appended file will contain the header row of each source file
}
\description{
Append all data files stored in the same folder + file name
}
