% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_hss_add_var.R
\name{C_hss_add_var}
\alias{C_hss_add_var}
\title{Generate a new HSS variable}
\usage{
C_hss_add_var(dat, name, location)
}
\arguments{
\item{dat}{HSS dataframe}

\item{name}{the name of the new variable (var_specific)}

\item{location}{the name of the variable that should come before the new var}
}
\value{
HSS dataframe which includes the new variable + a doc to be used for
the subsequent update of the dictionary
}
\description{
Update the dataset to include a new variable and generate a doc that will be
a necessary input in the following function which adds this new variable to
the dictionary. It is important to keep in mind that this will not change the
actual XLS form used to generate the dictionary. This is an internal
modification.
}
