% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_hss_clean_drop.R
\name{C_hss_clean_drop}
\alias{C_hss_clean_drop}
\alias{hss_filter_rows}
\alias{hss_select_cols}
\title{Drop HSS columns and rows}
\usage{
C_hss_clean_drop(dat)

hss_filter_rows(dat)

hss_select_cols(dat)
}
\arguments{
\item{dat}{The dataframe you want to remove entries from}
}
\value{
Returns a dataframe with invalid surveys removed as well as unnecessary columns.
}
\description{
This function is used to filter rows with invalid surveys from the HSS dataset.
Additionaly, columns containing notes & instructions to enumerators are removed
from the dataset.
}
