% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_hss_gen_prep.R
\name{C_hss_gen_prep}
\alias{C_hss_gen_prep}
\title{Coerce data into a convenient form for recoding}
\usage{
C_hss_gen_prep(dat, dict)
}
\arguments{
\item{dat}{The dataframe you want to remove entries from}

\item{dict}{Dictionary object}
}
\description{
This function is only a first step in a longer process. The goal here is to extract
-other- responses and display them in a convenient format that will allow for
later recoding using a different script. For step 2, use hss_rewrite.
}
\details{
This function will generate two spreadsheets. Only one will need modification (recode_1)
The second spreadsheet is generated with the purpose of erasing the _oth selection (recode_2)
}
