% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_hss_grow_dic.R
\name{C_hss_grow_dic}
\alias{C_hss_grow_dic}
\title{Update dictionary to include new variables}
\usage{
C_hss_grow_dic(dic)
}
\arguments{
\item{dic}{a dictionary object that is generated using the survey XLS form}
}
\value{
a modified dictionary object which includes the new variable generated
C_hss_add_var
}
\description{
This function takes the input produced by C_hss_add_var to enlarge the original
dictionary to include any variable generated during the cleaning/analysis process.
}
