% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_hss_NAtozero.R
\name{C_hss_NA2zero}
\alias{C_hss_NA2zero}
\title{Convert all Select Multi NA Values to Zero}
\usage{
C_hss_NA2zero(dat, num1, num2)
}
\arguments{
\item{dat}{dataframe to be used in conversion}

\item{num1}{coloumn number of first NA Multi variable}

\item{num2}{coloumn number of last NA Multi variable}
}
\value{
a dataframe that can be used for the relocation function
}
\description{
Convert all Select Multi NA Values to Zero
}
