% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_hss_NAgo.R
\name{C_hss_NAgo}
\alias{C_hss_NAgo}
\title{Convert Blank Cells to NA for Easy Splitting}
\usage{
C_hss_NAgo(dat)
}
\arguments{
\item{dat}{Data that needs to have blanks converted to NA}
}
\value{
Data that can be split
}
\description{
Convert Blank Cells to NA for Easy Splitting
}
