% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_hss_recode_check.R
\name{C_hss_recode_check}
\alias{C_hss_recode_check}
\title{Check if recode process is successful}
\usage{
C_hss_recode_check(dat, recode_file)
}
\arguments{
\item{dat}{Updated df that is generated after initial recode}

\item{recode_file}{Recode file to be used as reference for updates}
}
\value{
A update object (recode_format) with indicators on successful and failed matches
}
\description{
This function aims to review the updated DF (post-recode) to check if all values were replaced
correctly. The output will indicate successful changes, and NA where text was not matched.
}
