% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_hss_recode_fix.R
\name{C_hss_recode_fix}
\alias{C_hss_recode_fix}
\title{Fix recode mismatches through text replacement(s)}
\usage{
C_hss_recode_fix(recode_checked, prep)
}
\arguments{
\item{recode_checked}{recode_file generated after running the recode_check function}

\item{prep}{code_prep file generated when filtering out _oth text at start of process}
}
\value{
A new recode file that is 'fixed' and can be inputted to the rewrite function
}
\description{
This function is a follow-up to the recode_check function, and seeks to correct NA text
mismatches. The goal of this function is to generate an updated recode_file that can be shoved
back into the cleaning process. In premise, this function replaces broken recode text with
updated text that is recognized as part of the OG DF.
}
