% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_hss_relocate_multi.R
\name{C_hss_relocate_multi}
\alias{C_hss_relocate_multi}
\title{Relocate split_multiple variables to expected location}
\usage{
C_hss_relocate_multi(dat, output_1)
}
\arguments{
\item{dat}{data-frame with all variables (not ordered)}

\item{output_1}{data-frame list produced by C_hss_split_multi}
}
\value{
A data-frame that can be further cleaned
}
\description{
BASIC: Step two in the C_hss_split_multi pipleline. Variables are reordered
based on output provided by C_hss_split_multi, and relocated to their
expected location. Current determination on expected location relies on
previous experience with KOBO (where select_multiple variables were split),
and the reasonable expectation that sub-elements would come after their
_all parent.
PREP: Run C_hss_split_multi
NEXT: ...
}
