% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_hss_rewrite.R
\name{C_hss_rewrite}
\alias{C_hss_rewrite}
\title{Recode 'Other' Input}
\usage{
C_hss_rewrite(dat)
}
\arguments{
\item{dat}{actual survey data}
}
\description{
This function seeks to finalize the reoding procedure intialized by hss_gen_prep.
Key to this function is the construction of an xls sheet with the modified Var values.
This function is currently operational, but it requires modifications for clarity.
}
