% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_hss_space.R
\name{C_hss_space}
\alias{C_hss_space}
\title{Remove space(s) from text}
\usage{
C_hss_space(dat)
}
\arguments{
\item{dat}{df that needs recoding}
}
\value{
a df without spaces in the begining of text strings
}
\description{
This function is part of the recode pipeline and is necessary for the effective use of
the subsequent functions hss_gen_prep and hss_rewrite. The goal of this function is to eliminate
spaces at the start of a text string.
}
