% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_hss_split_multi.R
\name{C_hss_split_multi}
\alias{C_hss_split_multi}
\title{Split select_multiple survey variables}
\usage{
C_hss_split_multi(data, dict)
}
\arguments{
\item{data}{A dataframe that has been cleaned minimally - only C_hss_rename.}

\item{dict}{The dictionary reference that will be used to collect the
names associated with each variable.}
}
\value{
A dataframe that splits select_multiple variables into their components

A dataframe that will be used to reorder select_multiple variables
}
\description{
BASIC: Split select_multiple variables into their individual components,
with consideration for proper naming, conversation of 0s to NAs, and later
reordering of select_multiple components.
PREP: Run C_hss_create_dict, C_hss_rename
NEXT: Run C_hss_relocate_multi
}
