% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_hss_update_dict.R
\name{C_hss_update_dict}
\alias{C_hss_update_dict}
\title{Update your dictionary of HSS variable or value labels}
\usage{
C_hss_update_dict(dict, area)
}
\arguments{
\item{dict}{Dictionary object.}

\item{area}{The survey location for which the specific XLS form was made.}
}
\value{
A list containing two dataframes, with respectively variable- and value
names and their associated text labels *with added area filter.
}
\description{
This update is intended for better presentation of Iraq(subdistrict) data.
Note: Area refers to district in the XLS form, and in real terms to
subdistirct distribution.
}
