% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-SSDrop.R
\name{.drop_SS_A}
\alias{.drop_SS_A}
\title{Utility function for hss_drop_surveys - South Sudan
Drop invalid surveys based on selection criteria}
\usage{
.drop_SS_A(dat, audit)
}
\description{
HSS surveys are considered valid when they meet a number of selection criteria.
First, the respondent must give informed consent both at the beginning and the
end of the survey. Second, the survey should be excluded if the respondent seemed
uncomfortable and someone else interfered in the survey. Third, the survey must
meet some duration criteria: in general the survey should not be shorter than
22 minutes; for surveys with 3 or more incidents reported, it should not be shorter
than 27 minutes; for surveys with 4 or more incidents reported, it should not be
shorter than 32 minutes.
}
