% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_label.R
\name{.get_table_labels}
\alias{.get_table_labels}
\title{helper function for hss_label function. Don't run separately.}
\usage{
.get_table_labels(var, lang = "en")
}
\description{
helper function for hss_label function. Don't run separately.
}
\keyword{internal}
