% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_format.R
\name{.hss_format}
\alias{.hss_format}
\title{helper function for hss_format functions. Don't run separately.}
\usage{
.hss_format(table)
}
\description{
helper function for hss_format functions. Don't run separately.
}
\keyword{internal}
