% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vallabel_ar.R
\name{.vallabel_ar}
\alias{.vallabel_ar}
\title{Utility function for hss_vallabel 'Arabic'
Use a cumulative list of variable names to label their associated values}
\usage{
.vallabel_ar(data, dict)
}
\description{
The first half of this function uses hss_create_dic to produce 2 outputs
The first output is a DF with question labels, numeric, and r_names
The second output is a list which will be used to match var_names in the data set
to those in the DF
The result of this function is a new *labelled data set
}
