% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummydata.R
\docType{data}
\name{dummydata}
\alias{dummydata}
\title{Dummy HSS data}
\format{
A data frame with 437 rows and 21 variables:
\describe{
\item{gender}{gender of the respondent}
\item{county}{county of the respondent}
\item{hhsize}{Number of peeple in respondent's household}
\item{occupation}{Respondent's occupation}
\item{occupation_wagetype}{Respondent's type of wage}
\item{residence}{Has respondent lived in this location for the past 5 years}
\item{migr_nr}{Number of times respondent has moved in the past 5 years}
\item{migr_why_...}{List of binary variables of reasons for migration}
}
}
\usage{
dummydata
}
\description{
A dataset containing a number of common question types use by the HSS.
}
\keyword{datasets}
