% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_addlabels.R
\name{hss_addlabels}
\alias{hss_addlabels}
\title{Use a cumulative list of variable names to label their associated values}
\usage{
hss_addlabels(survey, data, location)
}
\arguments{
\item{survey}{the master xls form}

\item{data}{actual survey data set}

\item{location}{The location -directory- of the xls form}
}
\description{
The first half of this function uses hss_create_dic to produce 2 outputs
The first output is a DF with question labels, numeric, and r_names
The second output is a list which will be used to match var_names in the data set to those in the DF
The result of this function is a new *labelled data set
}
