% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_hss_clean_drop.R
\name{hss_clean_drop}
\alias{hss_clean_drop}
\alias{hss_filter_rows}
\alias{hss_select_cols}
\title{Drop HSS columns and rows}
\usage{
hss_clean_drop(dat)

hss_filter_rows(dat)

hss_select_cols(dat)
}
\arguments{
\item{dat}{The dataframe you want to remove entries from}

\item{prompt}{If TRUE, prompts user to provide criteria for removing test entries.
Input can be row numbers, a cut-off date, or device IDs. Accepts multiple criteria.
Set to FALSE to skip this prompt.}
}
\value{
Returns a dataframe with invalid surveys removed as well as unnecessary columns.
}
\description{
This function is used to filter rows with invalid surveys from the HSS dataset.
Additionaly, columns containing notes & instructions to enumerators are removed
from the dataset. The function prompts user for input to remove test rows. Set
\code{prompt} to FALSE to skip this step.
}
