% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_create_dict.R
\name{hss_create_dict}
\alias{hss_create_dict}
\title{Create a dictionary of HSS variable or value labels}
\usage{
hss_create_dict(form, type = "var")
}
\arguments{
\item{form}{Location of the XLS form}

\item{type}{Use "var" to create a dictionary of variable names and labels. Use "val" to create a dictionary of value names and labels.}
}
\value{
A dataframe containing variable or value names and their associated text labels.
}
\description{
Create a dictionary of HSS variable or value labels
}
\examples{
WIP
}
