% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_create_lookuplist.R
\name{hss_create_lookuplist}
\alias{hss_create_lookuplist}
\title{Create a lookup list of old & new variable names from renaming in a Stata .do file}
\usage{
hss_create_lookuplist(dofile)
}
\arguments{
\item{dofile}{Location of the Stata .do file with renaming commands}
}
\value{
A list with all old & new names of renamed variables
}
\description{
Create a lookup list of old & new variable names from renaming in a Stata .do file
}
\examples{
WIP
}
\seealso{
\code{\link[=namelookup]{namelookup()}} to lookup names in the generated list.
}
