% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_create_question_list.R
\name{hss_create_question_list}
\alias{hss_create_question_list}
\title{Create a list of questions from XLS form}
\usage{
hss_create_question_list(dict_path)
}
\arguments{
\item{dict_path}{Path to the XLS form}
}
\value{
A named character vector with question type as name and the question/variable name as value.
}
\description{
This function reads the XLS form and provides a list of all questions in the form as well as
the type of question.
}
