% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_drop_surveys_a.R
\name{hss_drop_surveys_a}
\alias{hss_drop_surveys_a}
\title{Drop invalid surveys based on selection criteria}
\usage{
hss_drop_surveys_a(dat, country)
}
\arguments{
\item{dat}{The dataframe containing HSS surveys}
}
\value{
A dataframe containing only those surveys that meet the selection criteria
}
\description{
HSS surveys are considered valid when they meet a number of selection criteria.
This is a combined function that allows for the selection of the survey country.
Additional details can be found within two other utility functions:
\itemize{
\item utils-IQDrop
\item utils-SSDrop
}
}
