% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_hss_export_formatted.R
\name{hss_export_formatted}
\alias{hss_export_formatted}
\title{Exports a formatted list of tables to .docx}
\usage{
hss_export_formatted(list, file = NULL)
}
\arguments{
\item{list}{A list containing flextable objects}

\item{file}{The file path to save to}
}
\value{
Outputs the selected flextable objects to a .docx file.
}
\description{
Exports the list of formatted tables created by \code{hss_write_formatted()} to a
.docx file in the specified location. One thing to keep in mind: the duration
of this process increase exponentially with the number of tables to write. Exporting
around 200 tables could take up to 10 minutes.
}
