% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_format.R
\name{hss_fomat_single}
\alias{hss_fomat_single}
\alias{hss_format_multi}
\alias{hss_format_group}
\title{Format tables using HSS style preferences}
\usage{
hss_fomat_single()

hss_format_multi()

hss_format_group()
}
\value{

}
\description{
Converts table to a flextable and applies formatting. Formattin preferences
are stored in \link{TBD}
}
