% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_hss_gen_prep.R
\name{hss_gen_prep}
\alias{hss_gen_prep}
\title{Coerce data into a convenient form for recoding}
\usage{
hss_gen_prep(dat)
}
\arguments{
\item{dat}{The dataframe you want to remove entries from}
}
\description{
This function is only a first step in a longer process. The goal here is to extract
-other- responses and display them in a convenient format that will allow for
later recoding using a different script. For step 2, use hss_recode.
}
