% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_get_multi.R
\name{hss_get_multi}
\alias{hss_get_multi}
\title{Find 'select multiple' response variables associated with a question variable.}
\usage{
hss_get_multi(var)
}
\arguments{
\item{var}{The variable for which 'select multiple' response variable are needed.}

\item{reverse}{If set to TRUE returns the new variable names. FALSE returns old names.}
}
\value{
a character string with all associated 'select multiple' variable names.
}
\description{
This function returns all associated answer variables with
a 'select-multiple' question variable. As of now, the only way to retrieve these variables
is by inputting the "_all" variable name.
}
