% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_label.R
\name{hss_label}
\alias{hss_label}
\title{Apply text labels to tables.}
\usage{
hss_label(table, var)
}
\arguments{
\item{table}{the table}

\item{var}{the variable name}
}
\value{

}
\description{
Apply text labels to tables.
}
