% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_loaddata.R
\name{hss_loaddata}
\alias{hss_loaddata}
\title{Load raw HSS Data}
\usage{
hss_loaddata(path, skip = 0)
}
\arguments{
\item{path}{Path to the HSS data file. Expects a .csv file}

\item{skip}{Number of rows to skip. Default is 0. Use this if you know how many rows contain test data.}
}
\value{
A dataframe with an added SurveyDuration column
}
\description{
Loads raw HSS datafile and performs some basic cleaning: proper encoding of Arabic, dates formatted as date-time objects, calculate survey duration.
}
