% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_lookup.R
\name{hss_lookup_var}
\alias{hss_lookup_var}
\alias{hss_lookup_val}
\alias{hss_lookup_list}
\title{Helper functions to look up variable or value names and labels.}
\usage{
hss_lookup_var(var, input_col, return_col)

hss_lookup_val(var, input_col, return_col)

hss_lookup_list(var, reverse = FALSE)
}
\arguments{
\item{var}{The variable name. Only a single character value is accepted, except for hss_lookup_list, which accepts a character string.}

\item{input_col}{The column in which the variable name can be found}

\item{return_col}{The column for which the associated value should be returned}

\item{reverse}{If set to TRUE will return the new name instead of the old name. Set to FALSE by default.}
}
\value{
A character string containing the found value(s)
}
\description{
These helper functions are used to look up info related to a
specific variable. This can be the variable label, the associated value labels
or the old/new variable name. These functions are mostly for use within other
functions.
}
