% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_merge_datasets.R
\name{hss_merge_datasets}
\alias{hss_merge_datasets}
\title{Merge HSS survey rounds into one dataframe}
\usage{
hss_merge_datasets(folder, mapping)
}
\arguments{
\item{folder}{a folder containing the .dta files to be merged. The function does
not check whether files belong to the same survey location.}

\item{mapping}{a mapping dataframe containing the variable names for the different
survey rounds and their relation to each other.}
}
\value{
Returns a dataframe consisting of all columns present in each of the
provided datasets.
}
\description{
Merges multiple survey rounds from the same location into a single dataframe.
The function expects stata .dta files and requires a mapping file containing
the changes of variable names throughout survey rounds. The resulting dataframe
contains only variables present in all datasets. Output is stripped of its haven
labels (if any).
}
