% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_mergetranslated.R
\name{hss_mergetranslated}
\alias{hss_mergetranslated}
\title{Merges HSS dataframe with dataframe containing translated columns.}
\usage{
hss_mergetranslated(df, df_translated)
}
\arguments{
\item{df}{The original HSS dataframe}

\item{df_translated}{The dataframe containing "_ar" and translated "_en" columns}
}
\value{
A merged dataframe containing HSS data and all translated columns.
}
\description{
Merge the translated columns with the original dataframe. All translated columns are
appended after their associated original column. Currently this only works for
Arabic -> English translation.
}
