% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_overview_multi.R
\name{hss_overview_multi}
\alias{hss_overview_multi}
\title{Get overview table for 'select-multiple' questions}
\usage{
hss_overview_multi(df, vars, percent = TRUE)
}
\arguments{
\item{df}{The dataframe containing the questions}

\item{vars}{Selected based on a string that identifies all relevant variables}

\item{percent}{T/F}
}
\value{

}
\description{
Get overview table for 'select-multiple' questions
}
