% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_overview_single.R
\name{hss_overview_single}
\alias{hss_overview_single}
\title{Create overview table for a group of similar 'select-one' questions}
\usage{
hss_overview_single(df, vars, percent = TRUE)
}
\arguments{
\item{df}{The dataframe containing the questions}

\item{vars}{Selected based on a string that identifies all relevant variables}

\item{percent}{T/F}
}
\value{
A dataframe with the selected questions.
}
\description{
Create overview table for a group of similar 'select-one' questions
}
