% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_rename.R
\name{hss_rename}
\alias{hss_rename}
\title{Rename HSS variables}
\usage{
hss_rename(df, dict_var, dict_val)
}
\arguments{
\item{df}{The dataframe containing the HSS survey questions}

\item{dict_var}{Dictionary object of variable names created from the survey's XLS form}

\item{dict_val}{Dictionary obect of value names created from the survey's XLS form.}
}
\value{
Returns a dataframe with names updated to new column names and non-question
columns removed.
}
\description{
This function is used to rename the variables in the HSS dataset from the old
Q... names to human-readable names. These new names are expected to be present
in the XLS form for the dataset (both the variable names, and the suffixes for
select-multiple answer options.). The function also removes any non-question columns
(notes etc.)
}
\seealso{
\code{hss_create_dict()}
}
