% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_table_overview.R
\name{hss_table_overview}
\alias{hss_table_overview}
\title{Create overview table}
\usage{
hss_table_overview(vars)
}
\arguments{
\item{vars}{a character vector with relevant variable strings.}
}
\value{
a dataframe
}
\description{
Create overview table
}
