% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_table_single.R
\name{hss_table_single}
\alias{hss_table_single}
\title{HSS Data Table Generation}
\usage{
hss_table_single(df, var, group, percent = TRUE)
}
\arguments{
\item{df}{A dataframe containing the variable of interest and grouping variable.}

\item{var}{A character string with the variable name of interest.}

\item{group}{A character string with the grouping (or disaggregation) variable.}

\item{percent}{Set to TRUE to show percentages. Set to FALSE to show counts.}
}
\value{
A contingency table with the variable of interest and grouping variable.
}
\description{
HSS Data Table Generation
}
