% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_hss_translate.R
\name{hss_translate}
\alias{hss_translate}
\title{Translate HSS content from Arabic to English}
\usage{
hss_translate(df, apikey)
}
\arguments{
\item{df}{the dataframe containing Arabic text columns. These columns are expected to end in "_ar"}
}
\value{
A dataframe containing Arabic text columns and their English translations.
#@seealso \code{hss_merge_translated}
}
\description{
Translate the content of Arabic text columns from the HSS dataset to English. This function
uses the translateR package to connect to Google Cloud Services. An API key from a GCS account is required.
Translated columns are returned as a separate dataframe. This makes it possible to export the translated
columns to e.g. check the accuracy of translations. The extracted and translated columns can be re-inserted
into the original dataset using the \code{hss_merge_translated} function.
}
