% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_hss_varlabel.R
\name{C_hss_VarLabel}
\alias{C_hss_VarLabel}
\title{Label all your named variables}
\usage{
C_hss_VarLabel(data, survey, dictionary)
}
\arguments{
\item{data}{actual survey data}

\item{survey}{the master xls form -character-}

\item{dictionary}{the dual list dictionary that should be generated early in the cleaning process}
}
\description{
This function generates labels for named variables. A prerequisite is ensuring that
variables have been already named using hss_rename.
}
