% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsschisq.R
\name{hsschisq}
\alias{hsschisq}
\title{Runs chi-squared test on selected variables}
\usage{
hsschisq(df, vars, group, full = FALSE)
}
\arguments{
\item{df}{The dataframe containing the variable(s) of interest}

\item{vars}{The variable(s) of interest. Accepts a single value or character string.}

\item{group}{The grouping (or disaggregation) variable.}

\item{full}{should the full results be returned. If set to FALSE, only p.value is returned}
}
\value{
A list or vector containing the results of the chi-squared test.
}
\description{
Runs chi-squared test on selected variables
}
\examples{
data <- testdata()
hsschisq(data, c("multi_option1", "multi_option2", "multi_option3"), "gender")
}
