% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hssdict_vars.R
\name{hssdict_vars}
\alias{hssdict_vars}
\alias{hssdict_vals}
\title{Create a dictionary of HSS variable and value labels}
\usage{
hssdict_vars(form)

hssdict_vals(form)
}
\arguments{
\item{form}{Location of the XLS form}
}
\value{
A dataframe containing variable and value names, and their associated English and Arabic labels.
}
\description{
Create a dictionary of HSS variable and value labels
}
\examples{
WIP
}
