% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hssformat.R
\name{hssformat}
\alias{hssformat}
\title{Apply HSS formatting to contingency table}
\usage{
hssformat(x)
}
\arguments{
\item{x}{A flextable object.}
}
\value{
a formatted flextable.
}
\description{
Apply HSS formatting to contingency table
}
\examples{
WIP
}
