% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hssformat_multi.R
\name{hssformat_multi}
\alias{hssformat_multi}
\title{Apply HSS formatting to multi-response contingecy table}
\usage{
hssformat_multi(x)
}
\arguments{
\item{x}{A flextable object.}
}
\value{
A flextable object.
}
\description{
Apply HSS formatting to multi-response contingecy table
}
\examples{
WIP
}
