% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hssfreqtable.R
\name{hssfreqtable}
\alias{hssfreqtable}
\title{HSS Data Table Generation}
\usage{
hssfreqtable(df, var, group, percent = TRUE)
}
\arguments{
\item{df}{A dataframe containing the variable of interest and grouping variable.}

\item{var}{A character string with the variable name of interest.}

\item{group}{A character string with the grouping (or disaggregation) variable.}

\item{percent}{Set to TRUE to show percentages. Set to FALSE to show counts.}
}
\value{
A contingency table with the variable of interest and grouping variable.
}
\description{
HSS Data Table Generation
}
\examples{
data <- testdata()
hssfreqtable(data, "singleresponse", "gender")
}
