% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hssfreqtable_sub.R
\name{hssfreqtable_sub}
\alias{hssfreqtable_sub}
\title{A specific implementation of hssfreqtable for sub-questions}
\usage{
hssfreqtable_sub(df, question, valid, var, group, percent = TRUE)
}
\arguments{
\item{df}{A dataframe containing the variable of interest and grouping variable.}

\item{question}{The variable containing the sub-setting question}

\item{valid}{A character string containing responses that should be included in the contingency table}

\item{var}{A character string with the variable name of interest.}

\item{group}{A character string with the grouping (or disaggregation) variable.}

\item{percent}{Set to TRUE to show percentages. Set to FALSE to show counts.}
}
\value{
A contingency table with the variable of interest and grouping variable.
}
\description{
A specific implementation of hssfreqtable for sub-questions
}
\examples{
data <- testdata()
hssfreqtable(data, "singleresponse", "gender")
}
