% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsslabel.R
\name{hsslabel}
\alias{hsslabel}
\alias{hsslabel_var}
\alias{hsslabel_val}
\title{Apply English or Arabic labels to a flextable object.}
\usage{
hsslabel(table, lang = "english", lookup = TRUE)

hsslabel_var(variable, lang = "english", lookup = TRUE)

hsslabel_val(variable, lang = "english", lookup = TRUE)
}
\arguments{
\item{table}{A flextable object.}

\item{lang}{The language to return. Accepts "en" or "ar".}

\item{lookup}{Set to TRUE to use a lookup table for renamed variables.}

\item{variable}{The variable name}

\item{dict_var}{The dictionary object containing the variable names and labels.}

\item{dict_val}{The dictionary object containing the value names and labels.}

\item{lookuplist}{If lookup is set to TRUE, provide object containing the lookup list.}
}
\value{
A flextable object with question and response labels.
}
\description{
Apply English or Arabic text labels  to a flextable object.
hsslabel() calls upon hsslabel_var() and hsslabel_val() which return a
character string of length 1 containing the variable label, and a character
string of length n containing the value labels respectively.
}
\examples{
WIP
}
\seealso{
\code{\link[=hssdict]{hssdict()}} to create a variable or value label dictionary, \code{\link[=hssnewnames]{hssnewnames()}} for creating a lookuplist of renamed variables.
}
