% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hssmultitable.R
\name{hssmultitable}
\alias{hssmultitable}
\title{Generate contingency table for multiresponse questions}
\usage{
hssmultitable(df, resp, group, percent = TRUE)
}
\arguments{
\item{df}{The dataframe containing the multiresponse questions}

\item{resp}{A character string of all response variables to include}

\item{group}{A grouping (or disaggregation) variable.}

\item{percent}{Set to TRUE to show percentages. Set to FALSE to show counts}
}
\value{
A contingency table containing the multiresponse answers and a grouping variable
}
\description{
Generate contingency table for multiresponse questions
}
\examples{
data <- testdata()
hssmultitable(data, c("multi_option1", "multi_option2", "multi_option3"), "gender")
}
