% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hssmultitable_sub.R
\name{hssmultitable_sub}
\alias{hssmultitable_sub}
\title{A specific implementation of hssmultitable for sub-questions}
\usage{
hssmultitable_sub(df, question, valid, resp, group, percent = TRUE)
}
\arguments{
\item{df}{The dataframe containing the multiresponse questions}

\item{question}{The variable containing the sub-setting question}

\item{valid}{A character string containing responses that should be included in the contingency table}

\item{resp}{A character string of all response variables to include}

\item{group}{A grouping (or disaggregation) variable.}

\item{percent}{Set to TRUE to show percentages. Set to FALSE to show counts}
}
\value{

}
\description{
A specific implementation of hssmultitable for sub-questions
}
\examples{
data <- testdata()
hssmultitable_sub(
  df = data,
  question = "multiresponse",
  valid = "Yes",
  resp = c("multi_option1", "multi_option2", "multi_option3"),
  group = "location",
  percent = TRUE
)
}
