% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hssnamelookup.R
\name{hssnamelookup}
\alias{hssnamelookup}
\title{Lookup old variable name by providing the new one.}
\usage{
hssnamelookup(list, lookup, reverse = FALSE)
}
\arguments{
\item{list}{The list containing the lookup value}

\item{lookup}{The old variable name}

\item{reverse}{Set to TRUE to lookup new variable name by providing the old one.}
}
\value{
A character string
}
\description{
Lookup old variable name by providing the new one.
}
\examples{
WIP
}
\seealso{
\code{\link[=hssnewnames]{hssnewnames()}} to create a lookup list
}
