% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsstable_multi.R
\name{hsstable_multi}
\alias{hsstable_multi}
\alias{hsstable_multi_sub}
\title{Generate contingency table for multiresponse questions}
\usage{
hsstable_multi(df, resp, group, percent = TRUE)

hsstable_multi_sub(df, question, valid, resp, group, percent = TRUE)
}
\arguments{
\item{df}{The dataframe containing the multiresponse questions}

\item{resp}{A character string of all response variables to include}

\item{group}{A grouping (or disaggregation) variable.}

\item{percent}{Set to TRUE to show percentages. Set to FALSE to show counts}

\item{question}{The variable containing the sub-setting question}

\item{valid}{A character string containing responses that should be included in the contingency table}
}
\value{
A contingency table containing the multiresponse answers and a grouping variable
}
\description{
Generate contingency table for multiresponse questions
}
\examples{
data <- testdata()
hssmultitable(data, c("multi_option1", "multi_option2", "multi_option3"), "gender")
}
