% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsstable_single.R
\name{hsstable_single}
\alias{hsstable_single}
\alias{hsstable_single_sub}
\title{HSS Data Table Generation}
\usage{
hsstable_single(df, var, group, percent = TRUE)

hsstable_single_sub(df, question, valid, var, group, percent = TRUE)
}
\arguments{
\item{df}{A dataframe containing the variable of interest and grouping variable.}

\item{var}{A character string with the variable name of interest.}

\item{group}{A character string with the grouping (or disaggregation) variable.}

\item{percent}{Set to TRUE to show percentages. Set to FALSE to show counts.}

\item{question}{The variable containing the sub-setting question}

\item{valid}{A character string containing responses that should be included in the contingency table}
}
\value{
A contingency table with the variable of interest and grouping variable.
}
\description{
HSS Data Table Generation
}
\examples{
data <- testdata()
hssfreqtable(data, "singleresponse", "gender")
}
