% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_hss_act_perf.R
\name{R_hss_act_perf}
\alias{R_hss_act_perf}
\title{Reports & Publications: Actor Performance}
\usage{
R_hss_act_perf(data, actor)
}
\arguments{
\item{data}{HSS dataframe}

\item{actor}{variable name relating to actor performance (_Perf)}
}
\value{
A breakdown of actor performance
}
\description{
This function calculates actor performance.
Performance is displayed in console and as a simplified plot.
This function does not require an entity.
}
