% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_hss_age_count.R
\name{R_hss_age_count}
\alias{R_hss_age_count}
\title{Reports & Publications: Age Distriubution}
\usage{
R_hss_age_count(dat)
}
\arguments{
\item{dat}{HSS dataframe}
}
\value{
age breakdown for data
}
\description{
Reports & Publications: Age Distriubution
}
