% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_hss_statement_test.R
\name{R_hss_statement_test}
\alias{R_hss_statement_test}
\title{Reports & Publications: Statement Preference (A/B)}
\usage{
R_hss_statement_test(data, var)
}
\arguments{
\item{data}{HSS dataframe}

\item{var}{variable name of statement}
}
\value{
a printed summary
}
\description{
This function quickly summarizes respondent preferences for a statement (either A or B).
Input for the statement is the R_name of the variable, e.g st_securityforces.
Function does not require entity to run.
}
