% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_hss_waffle_sec.R
\name{R_hss_waffle_sec}
\alias{R_hss_waffle_sec}
\title{Reports & Publications: Waffle visualization of security}
\usage{
R_hss_waffle_sec(data, x)
}
\arguments{
\item{data}{dataframe used for visualization}

\item{x}{number of waffle squares}
}
\value{
a file containing the visualization
}
\description{
Reports & Publications: Waffle visualization of security
}
