% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-split.R
\name{split_2}
\alias{split_2}
\title{Split select_multiple survey variables}
\usage{
split_2(data, dict)
}
\description{
BASIC: Split select_multiple variables into their individual components,
with consideration for proper naming, conversation of 0s to NAs, and later
reordering of select_multiple components.
}
