% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_hss_export_tables.R
\name{T_hss_export_tables}
\alias{T_hss_export_tables}
\title{Exports a list of data tables to CSV}
\usage{
T_hss_export_tables(df_list, questions, path)
}
\arguments{
\item{df_list}{A named list containing the data tables. The output of hss_write_tables() usually.}

\item{path}{Where to store the .csv output.}
}
\value{
A CSV file stored at the provided location.
}
\description{
Exports a list of HSS tables created with \code{hss_write_tables()} to a CSV file in
the specified location. Note that this function uses sink() to write multiple
tables to the same file.
}
\seealso{
hss_write_tables()
}
