% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_hss_label.R
\name{T_hss_label}
\alias{T_hss_label}
\title{Apply text labels to tables.}
\usage{
T_hss_label(table, var, grouping, dict, lang = "en")
}
\arguments{
\item{table}{the table for which labels should be applied}

\item{var}{the variable name. Used to determine the question label and to look up the appropriate
response labels.}

\item{grouping}{the grouping variable. Will be used to apply appropriate column headers}

\item{dict}{The variable/value dictionary object}

\item{language}{To determine the language of labels to be applied.}
}
\value{
A flextable object with the original table values and appropriate question & response labels.
}
\description{
Apply text labels to the specified table. Labels are taken from the dictionaries
created from the XLS forms. This converts the table to a flextable object. Works for EN or AR text labels.
If used with AR text labels some optional formatting is applied to correctly display text.
}
