% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_hss_write_tables.R
\name{T_hss_write_tables}
\alias{T_hss_write_tables}
\title{Write tables to a list for selected variables}
\usage{
T_hss_write_tables(df, questions, group, percent = TRUE, digits = 3)
}
\arguments{
\item{df}{The dataframe containing relevant variables}

\item{questions}{A named character vector containing variable names and the table type required.}

\item{group}{The desired grouping/disaggregation variable.}

\item{percent}{Logical vector if the table should be created with percentage values. If set to FALSE, counts are shown.}
}
\value{
A named list of dataframes containing contingecy tables for the specified
questions of the HSS.
}
\description{
Write (unformatted) tables to a list for all variable/question names contained in a
questions list. A question list for all questions in the current survey can be
made  with the \code{hss_create_question_list} function. This function silently skips
any questions that cause errors during creation of the contingency table, and
returns a list item of length NULL instead. This is useful to automatically omit
any question variables that have been removed during data cleaning, but might in
some cases lead to unexpected behavior. The list of dataframes created with this
function can be used as input for \code{hss_export_tables} to write the tables to a
.csv file.
}
\seealso{
\code{hss_export_tables}, \code{hss_create_question_list}
}
