% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_hss_read_dat.R
\name{C_hss_read_dat}
\alias{C_hss_read_dat}
\title{Read Encoded Iraq (Arabic) Data}
\usage{
C_hss_read_dat()
}
\value{
Raw data with proper encoding for Arabic text
}
\description{
This function cleans up the procedure for reading-in Iraq (Arabic) data.
Encoding is a key faccet of this function. Arabic letters are not
legibile without this function.
}
