% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_hss_relocate_multi.R
\name{C_hss_relocate_multi}
\alias{C_hss_relocate_multi}
\title{Relocate split_multiple variables to expected location}
\usage{
C_hss_relocate_multi(data, dict, var_names_2, output_multi)
}
\arguments{
\item{var_names_2}{naming list to assist in relocation process}

\item{output_multi}{data-frame list produced by C_hss_split_multi}

\item{dat}{data-frame with all variables (not ordered)}
}
\value{
A data-frame that can be further cleaned
}
\description{
Relocate split_multiple variables to expected location
}
