% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_addlabels.R
\name{hss_addlabels}
\alias{hss_addlabels}
\title{Use a list of variable names to label their associated values
The first half of this function uses hss_create_dic to produce 2 outputs
The first output is a DF with question labels, numerics, and r_names
The second output is a list which will be used to match var_names in the dataset to those in the DF
The result of this function is a new *labelled dataset}
\usage{
hss_addlabels(survey, data)
}
\arguments{
\item{survey}{the master xls form}

\item{data}{actual survey dataset}
}
\description{
Use a list of variable names to label their associated values
The first half of this function uses hss_create_dic to produce 2 outputs
The first output is a DF with question labels, numerics, and r_names
The second output is a list which will be used to match var_names in the dataset to those in the DF
The result of this function is a new *labelled dataset
}
