% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_chisq.R
\name{hss_chisq}
\alias{hss_chisq}
\title{Runs chi-squared test on selected variables}
\usage{
hss_chisq(df, vars, group, full = FALSE)
}
\arguments{
\item{df}{The dataframe containing the variable(s) of interest}

\item{vars}{The variable(s) of interest. Accepts a single value or character string.}

\item{group}{The grouping (or disaggregation) variable.}

\item{full}{should the full results be returned. If set to FALSE, only p.value is returned}
}
\value{
A vector containing the results of the chi-squared test for the selected
variables. If full is set to TRUE this will be a list, otherwise an atomic numeric vector.
}
\description{
This functions runs chi-squared significance tests for one or more variables
in a dataframe with a single cross-variable. Variable names can be passed as a
single object or as a character string.
}
\examples{
# Generate dummy data
df <- hss_dummydata()

# Calculte p-value for chi-squared test on a 'select-one' question.
pval_single <- hss_chisq(test, "singleresponse", "gender")

# Calculate p-value for chi-squared test on a 'select-multiple' question.
multi_options <- c("multi_option1", "multi_option2", "multi_option3")
pval_multi <- hss_chisq(test, multi_options, "gender")

# Chi-squared test with full output
full <- hss_chisq(f, "singleresponse", full = TRUE)
}
