% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_chisq.R
\name{hss_chisq}
\alias{hss_chisq}
\title{Runs chi-squared test on selected variables}
\usage{
hss_chisq(df, vars, group, full = FALSE)
}
\arguments{
\item{df}{The dataframe containing the variable(s) of interest}

\item{vars}{The variable(s) of interest. Accepts a single value or character string.}

\item{group}{The grouping (or disaggregation) variable.}

\item{full}{should the full results be returned. If set to FALSE, only p.value is returned}
}
\value{
A vector containing the results of the chi-squared test for the selected variables.
}
\description{
Runs chi-squared test on selected variables
}
