% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_chisq.R
\name{hss_chisq}
\alias{hss_chisq}
\alias{hss_chisq_formatted}
\title{Runs chi-squared test on selected variables}
\usage{
hss_chisq(df, var, group, full = FALSE, multi = FALSE)

hss_chisq_formatted(df, var, group)
}
\arguments{
\item{df}{The dataframe containing the variable(s) of interest}

\item{var}{The names of the variable(s) of interest. Accepts a single value or a character vector
containing multiple variable names.}

\item{group}{The grouping (or disaggregation) variable.}

\item{full}{TRUE/FALSE if the full results should be returned. Set to FALSE by
default, returning only the p.value.}

\item{multi}{Set to FALSE if used for a 'select-one' question. Set to TRUE if used for a
'select-multiple question to look up the corresponding response options.}
}
\value{
A vector containing the results of the chi-squared test for the selected
variables. If full is set to TRUE this will be a list, otherwise an atomic numeric vector.
For hss_chisq_formatted the output is a character vector.
}
\description{
This functions runs chi-squared significance tests for one or more variables
in a dataframe, with a single cross-variable. Variable names are passed as a
character vector containing one or more variable names. hss_chisq_formatted
returns a formatted character string containing a verbose description of the p-value.
This verbose output is used in formatted data tables.
}
\examples{
# Create dummy dictionary
dict_var <- dummy_var
dict_val <- dummy_val
# Calculte p-value for chi-squared test on a 'select-one' question.
hss_chisq(dummydata, "migr_nr", "gender")

# Calculate p-value for chi-squared test on a 'select-multiple' question.
hss_chisq(dummydata, "migr_why_all", "gender", multi = TRUE)

# Chi-squared test with full output
hss_chisq(dummydata, "migr_nr", full = TRUE)

# Formatted output
hss_chisq_formatted(dummydata, "migr_nr", "gender")
}
