% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_combine_questions.R
\name{hss_combine_single}
\alias{hss_combine_single}
\alias{hss_combine_multi}
\title{Combine multiple questions with the same response options}
\usage{
hss_combine_single(df, var)

hss_combine_multi(df, var)
}
\arguments{
\item{df}{The dataframe containing the questions}

\item{var}{The shared element in all of the question/variable names}
}
\value{
This returns a dataframe with the response options and the number of responses for that option.
}
\description{
The HSS data contains several questions that are repeated a number of times (for
example to ask the same question for several different actors/topics). These
questions can be represented with individual tables for each questions, but for some
publications it might be preferable to combine these questions in a single overview table.
}
