% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_create_dict.R
\name{hss_create_dict}
\alias{hss_create_dict}
\title{Create a dictionary of HSS variable or value labels}
\usage{
hss_create_dict(form, type, location = NULL)
}
\arguments{
\item{form}{Path to the XLS form.}

\item{type}{Use "var" to create a dictionary of variable names and labels.
Use "val" to create a dictionary of value names and labels.}

\item{location}{For variable dictionaries, filters to the specified location to ensure that
proper county/district labels are available.}
}
\value{
A dataframe containing variable or value names and their associated text labels.
}
\description{
Create a dictionary from an XLS form. The dictionary is stored as a dataframe,
and contains the contents of the XLS form stored in a (slightly modified) format
that allows it to be used as a dictionary for looking up variable & value labels,
This function needs to be run separately for variable- and value dictionaries.
}
