% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_hss_create_dict.R
\name{hss_create_dict}
\alias{hss_create_dict}
\title{Create a dictionary of HSS variable or value labels}
\usage{
hss_create_dict(form, location = NULL)
}
\arguments{
\item{form}{Path to the XLS form.}

\item{location}{The survey location for which the specific XLS form was made.}
}
\value{
A list containing two dataframes, with respectively variable- and value
names and their associated text labels.
}
\description{
Create a dictionary from an XLS form. The dictionary is stored as a dataframe,
and contains the contents of the XLS form stored in a (slightly modified) format
that allows it to be used as a dictionary for looking up variable & value labels.
This second argument of this function allows for filtering of questions and responses
that are specific to, or different between, certain locations. Please check the
XLS form in question for survey location abbreviations.
}
