% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_export_formatted.R
\name{hss_export_formatted}
\alias{hss_export_formatted}
\title{Exports a formatted list of tables to .docx}
\usage{
hss_export_formatted(list, path = NULL)
}
\arguments{
\item{list}{A list containing flextable objects}

\item{path}{The file path to save to}
}
\value{

}
\description{
Exports a formatted list of tables to .docx
}
