% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_export_tables.R
\name{hss_export_tables}
\alias{hss_export_tables}
\title{Exports a list of data tables to .csv}
\usage{
hss_export_tables(df_list, path)
}
\arguments{
\item{df_list}{A named list containing the data tables. The output of hss_write_tables() usually.}

\item{path}{Where to store the .csv output.}
}
\value{

}
\description{
Exports a list of data tables to .csv
}
