% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_format.R
\name{hss_format_single}
\alias{hss_format_single}
\alias{hss_format_multi}
\alias{hss_format_group}
\title{Format tables using HSS style preferences}
\usage{
hss_format_single(table)

hss_format_multi(table)

hss_format_group()
}
\value{
A flextable object with formatting applied.
}
\description{
Converts table to a flextable and applies formatting, according to set preferences. There are two
separate functions for single-response and multiple-response question tables. Neither one needs any
further arguments.
}
