% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_hss_load_audit.R
\name{hss_load_audit}
\alias{hss_load_audit}
\alias{hss_clean_audit}
\title{Load audit files for HSS survey}
\usage{
hss_load_audit(path)

hss_clean_audit(dat)
}
\arguments{
\item{path}{Path to a folder containing the audit .csv files}

\item{dat}{The resulting dataframe from hss_load_audit}
}
\description{
These functions are used to load individual audit files for a HSS survey as a
single dataframe and apply some basic cleaning (currently only adds a new column
containing the duration spend on a screen.
}
