% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_hss_overview_multi.R
\name{hss_overview_multi}
\alias{hss_overview_multi}
\title{Create overview table for 'select-multiple' questions}
\usage{
hss_overview_multi(df, vars, percent = TRUE)
}
\arguments{
\item{df}{The dataframe containing the questions.}

\item{vars}{A commom character string that is shared between all relevant variables.}

\item{percent}{Set to TRUE to display percentages, set to FALSE to display counts. Default is TRUE}
}
\value{
A dataframe with responses for the selected variables.
}
\description{
This creates overview tables for multiple questions that contain the same response options.
For example when the same question is repeated for different events. This implementation is specifically for
'select-multiple' questions. For 'select-one' questions see \code{\link{hss_overview_single}}
}
