% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_overview_single.R
\name{hss_overview_single}
\alias{hss_overview_single}
\title{Create overview table for a group of similar 'select-one' questions}
\usage{
hss_overview_single(df, vars, percent = TRUE)
}
\arguments{
\item{df}{The dataframe containing the questions}

\item{vars}{A commom character string that is shared between all relevant variables.}

\item{percent}{Set to TRUE to display percentages, set to FALSE to display counts. Default is TRUE}
}
\value{
A dataframe with responses for the selected variables.
}
\description{
This creates overview tables for multiple questions that contain the same response options.
For example when the same question is repeated for different events. This implementation is specifically for
'select-one' questions. For 'select-multiple' questions see \code{\link{hss_overview_multi}}
}
