% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_surveyduration.R
\name{hss_surveyduration}
\alias{hss_surveyduration}
\title{Convert columns to date-time and calculate survey duration}
\usage{
hss_surveyduration(df)
}
\arguments{
\item{df}{A dataframe containing the Kobo survey results}
}
\value{
A dataframe with columns 'SubmissionDate', 'end' and 'start' converted to ymd_hms format; and a new column SurveyDuration as a duration in seconds.
}
\description{
Convert columns to date-time and calculate survey duration
}
\examples{
WIP
}
