% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_surveyduration.R
\name{hss_surveyduration}
\alias{hss_surveyduration}
\title{Calculate survey duration}
\usage{
hss_surveyduration(data)
}
\arguments{
\item{data}{The most recent iteration of the data set}
}
\value{
A dataframe with an added SurveyDuration column
}
\description{
Loads raw HSS datafile and performs some basic cleaning: proper encoding of
Arabic, dates formatted as date-time objects, calculate survey duration.
}
