% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_translate.R
\name{hss_translate}
\alias{hss_translate}
\title{Translate HSS content from Arabic to English}
\usage{
hss_translate(df)
}
\arguments{
\item{df}{the dataframe containing Arabic text columns. These columns are expected to end in "_ar"}
}
\value{
A dataframe containing Arabic text columns and their English translations.
}
\description{
Translate HSS content from Arabic to English
}
