% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_VarLabel.R
\name{hss_VarLabel}
\alias{hss_VarLabel}
\title{Label all your named variables}
\usage{
hss_VarLabel(data, survey, location)
}
\arguments{
\item{data}{actual survey data}

\item{survey}{the master xls form -character-}

\item{location}{the location of the master xls form -path-}
}
\description{
This function generates labels for named variables. A prerequisite is ensuring that
variables have been already named using hss_rename.
}
