% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_write_tables.R
\name{hss_write_tables}
\alias{hss_write_tables}
\title{Write tables to a list for selected variables}
\usage{
hss_write_tables(df, questions, group, percent = TRUE)
}
\arguments{
\item{df}{The dataframe containing relevant variables}

\item{questions}{A named character vector containing variable names and the table type required.}

\item{group}{The desired grouping/disaggregation variable.}

\item{percent}{Logical vector if the table should be created with percentage values. If set to FALSE, counts are shown.}
}
\value{
A named list of dataframes.
}
\description{
Write tables to a list for selected variables
}
