% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hssdict.R
\name{hssdict}
\alias{hssdict}
\title{Create a dictionary of HSS variable or value labels}
\usage{
hssdict(form, type = "vars")
}
\arguments{
\item{form}{Location of the XLS form}

\item{type}{Use "vars" to create a dictionary of variable names and labels. Use "vals" to create a dictionary of value names and labels.}
}
\value{
A dataframe containing variable or value names, and their associated English and Arabic labels.
}
\description{
Create a dictionary of HSS variable or value labels
}
\examples{
WIP
}
