% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hssmultiresponses.R
\name{hssmultiresponses}
\alias{hssmultiresponses}
\title{Find response options for 'select multiple' questions}
\usage{
hssmultiresponses(variable, dict_var, dict_val, lookuplist, lookup = FALSE)
}
\arguments{
\item{variable}{The question for which response options should be returned. This is the "Q" number or, if lookup is set to TRUE, this is the new variable name.}

\item{dict_var}{The dictionary object containing the variable names and labels.}

\item{dict_val}{The dictionary object containing the value names and labels.}

\item{lookuplist}{The object containing the lookup list.}

\item{lookup}{Set to TRUE to use a lookup table for renamed variables.}
}
\value{

}
\description{
Find response options for 'select multiple' questions
}
\examples{
WIP
}
