% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_hss_age_count.R
\name{R_hss_age_count}
\alias{R_hss_age_count}
\title{Reports & Publications: Age Distribution}
\usage{
R_hss_age_count(dat)
}
\arguments{
\item{dat}{HSS dataframe}
}
\value{
age breakdown for data
}
\description{
This function provides an easy overview of the survey age distribution.
Future edits may require changing the age range.
}
