% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_hss_sum_values.R
\name{R_hss_sum_values}
\alias{R_hss_sum_values}
\title{Calculate variable sums based on group}
\usage{
R_hss_sum_values(dat, dictionary, group)
}
\arguments{
\item{dat}{clean data-set}

\item{dictionary}{dictionary df/list used as a reference}

\item{group}{variable shorthand used to select utility function}
}
\value{
Text/Table providing required calculations for HSS outputs
}
\description{
This multipurpose function provides a selection of utility functions which can
be used to calculate breakdowns and sums necessary for HSS publications. In case of
any issues, please input '?' into the group parameter to list options.
}
